requirements.txt
pandas
python-dotenv
openai
anthropic
google-generative-ai